+++ app/services/menus/adopt_from_public_catalog.rb
@@ -25,7 +25,7 @@ module Menus
 
       ApplicationRecord.transaction do
         recipe_map = {}
-        source.menu_entries.where.not(recipe_id: nil).distinct.pluck(:recipe_id).compact.each do |rid|
+        source.menu_entries.where.not(dish_id: nil).distinct.pluck(:dish_id).compact.each do |rid|
           src_recipe = Recipe.find(rid)
           recipe_map[rid] = DuplicateRecipeForAdopter.call(source_recipe: src_recipe, adopter: adopter).id
         end